Add clean-up for offline cats
Reasoning:
When we do cats without the offline rootfses, we create our release
names with a timestamp suffix. See bosh-cr task
But, for cats with offline rootfses, We use the `create-release` script
inside the offline rootfs repo. This under the hood calls
`bosh create-release` without the timestamp name suffix.
Because of this, when concourse uploads the compiled release
it's a duplicate and it errors out.

- [#167233672]

Signed-off-by: Arjun Sreedharan <asreedharan@pivotal.io>